style(import-script): rename files
